refactor(idea_service): extract hierarchy + rollup ops (#163)
  before: 1007
  after:  870 (-137)
